fix streling module and update test
