+++ sterling/transfer.py
@@ -1,13 +1,15 @@
+import json
 from urllib.parse import urljoin
 from url import url as BASE_URL
 from sterling.Sterling import Sterling
 from sterling.utils.calls import calls
 
 
+
 class Transfer(Sterling):
-    def InterbankNameEnquiry(self, query):
+    def InterbankTransferReq(self, body):
         url = urljoin(BASE_URL(self.url),
-                      "/sterling/TransferAPIs/api/Spay/InterbankNameEnquiry")
+                      "/sterling/accountapi/api/Spay/InterbankTransferReq")
         header = self.headers
-        apiResponse = calls("get", url, query, "", header)
-        return apiResponse
+        apiResponse = calls("post", url, None, json.dumps(body), header)
+        return (apiResponse)
